backend/lambdas/summarize/requirements.txt
boto3==1.34.51
PyPDF2==3.0.1
python-docx==1.1.0
